fix: move schema to app
